[buildqt5] Add QT_SOURCE_PACKAGE variable
Added variable QT_SOURCE_PACKAGE for specifying the name of the Qt
source package to build. The default value is
“qt-everywhere-opensource-src-5.2.1”.

The primary use for this variable is to make it simpler to build other
versions of Qt on platforms where the default version does not compile
out of the box (such as OS X 10.10, which requires Qt 5.4.0 or later).