Enable SSL support in installers on Windows
This is done by statically linking OpenSSL binaries.

OpenSSL binaries must be installed on the build host. The recommentation
found on Qt wiki [1] is to install binaries available on this site [2].

On Linux it is enough to install libssl-dev on build host, on Mac
everything seems to be in place by default.

[1] https://wiki.qt.io/Technical_FAQ#How_can_I_make_the_QSsl..._classes_work_on_Windows.3F
[2] https://slproweb.com/products/Win32OpenSSL.html