buildqtc: Disable debug_and_release on Windows
This option is enabled by default (just) on Windows. The effect is that
each project file is processed three times to produce Makefiles for both
debug and release builds, release being the default. With Qt Creator
sources this causes project warnings when the project files are parsed
for the -- never actually used -- debug mode.